updated to more correct BIGINT in db

+++ apl-core/src/main/java/com/apollocurrency/aplwallet/apl/core/db/dao/factory/ShardStateFactory.java
@@ -15,7 +15,7 @@ public class ShardStateFactory extends AbstractArgumentFactory<ShardState> {
 
 
     public ShardStateFactory() {
-        super(Types.INTEGER);
+        super(Types.BIGINT);
     }
 
     @Override
